feat: add account permission check

+++ pkg/accounts/service/avatar.ts
@@ -23,13 +23,16 @@ export class AccountAvatarService {
    * avatar specification:
    * - NSFW Media can't be used as avatar image.
    * - Media must be image type(ToDo).
+   * - Media author must be actor.
    *
    * @param accountID
    * @param mediumID
+   * @param actorID
    */
   async create(
     accountID: AccountID,
     mediumID: MediumID,
+    actorID: AccountID,
   ): Promise<Result.Result<Error, void>> {
     const mediumRes = await this.mediaModule.fetchMedia(mediumID);
     if (Result.isErr(mediumRes)) {
@@ -46,6 +49,14 @@ export class AccountAvatarService {
     }
     // ToDo: Check media type
 
+    const isAllowedRes = this.isAllowed('set', actorID, {
+      targetAccount: accountID,
+      medium,
+    });
+    if (Result.isErr(isAllowedRes)) {
+      return isAllowedRes;
+    }
+
     const res = await this.avatarRepository.create(accountID, mediumID);
     if (Result.isErr(res)) {
       return res;
@@ -57,8 +68,19 @@ export class AccountAvatarService {
   /**
    * @description Unset account avatar image.
    * @param accountID
+   * @param actorID
    */
-  async delete(accountID: AccountID): Promise<Result.Result<Error, void>> {
+  async delete(
+    accountID: AccountID,
+    actorID: AccountID,
+  ): Promise<Result.Result<Error, void>> {
+    const isAllowedRes = this.isAllowed('unset', actorID, {
+      targetAccount: accountID,
+    });
+    if (Result.isErr(isAllowedRes)) {
+      return isAllowedRes;
+    }
+
     const res = await this.avatarRepository.delete(accountID);
     if (Result.isErr(res)) {
       return res;
@@ -75,6 +97,47 @@ export class AccountAvatarService {
     accountID: AccountID,
   ): Promise<Result.Result<Error, Medium>> {
     return await this.avatarRepository.findByID(accountID);
+  }
+
+  private isAllowed(
+    action: 'set' | 'unset',
+    actor: AccountID,
+    resources: { targetAccount: AccountID; medium?: Medium },
+  ): Result.Result<Error, void> {
+    switch (action) {
+      case 'set':
+        // NOTE: actor must be same as target.
+        if (actor !== resources.targetAccount) {
+          return Result.err(
+            new AccountInsufficientPermissionError(
+              'Actor must be same as target',
+              { cause: null },
+            ),
+          );
+        }
+        // NOTE: media author must be actor.
+        if (resources.medium?.getAuthorId() !== actor) {
+          return Result.err(
+            new AccountInsufficientPermissionError(
+              'Media author must be actor',
+              { cause: null },
+            ),
+          );
+        }
+        return Result.ok(undefined);
+      case 'unset':
+        // NOTE: actor must be same as target.
+        if (actor !== resources.targetAccount) {
+          return Result.err(
+            new AccountInsufficientPermissionError(
+              'Actor must be same as target',
+              { cause: null },
+            ),
+          );
+        }
+
+        return Result.ok(undefined);
+    }
   }
 }
 export const accountAvatarSymbol = Ether.newEtherSymbol<AccountAvatarService>();
